fix: replace Typescript enum with JS object
It allows users to use directly "directory" of "file" as filter instead
of importing the FileType enum.
Typescript documentation also recommends to use JS objects instead of
enum when `as const` could suffice.
I also use the sources in tests instead of the dist folder. I'm not
sure why, Ava does not recognize the module members when running even
if the library seems to work as expected when testing in an external
project.